requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq